realEstate/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-phone-field==1.8.1
djangorestframework==3.11.0
feedparser==5.2.1
idna==2.9
newsapi==0.1.1
newsapi-python==0.2.6
oauthlib==3.1.0
Pillow==7.1.2
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
